Move RSA key to where it's used
